move test setup init to more appropriate file
