+++ units/ftdc_rollups_test.go
@@ -5,42 +5,15 @@ import (
 	"os"
 	"runtime"
 	"testing"
-	"time"
 
 	"github.com/evergreen-ci/cedar"
 	"github.com/evergreen-ci/cedar/model"
 	"github.com/evergreen-ci/cedar/perf"
 	"github.com/mongodb/amboy/queue"
-	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.mongodb.org/mongo-driver/bson"
 )
-
-const testDBName = "cedar_test_config"
-
-func init() {
-	env, err := cedar.NewEnvironment(context.Background(), testDBName, &cedar.Configuration{
-		MongoDBURI:    "mongodb://localhost:27017",
-		DatabaseName:  testDBName,
-		SocketTimeout: time.Minute,
-		NumWorkers:    2,
-	})
-	if err != nil {
-		panic(err)
-	}
-
-	cedar.SetEnvironment(env)
-}
-
-func tearDownEnv(env cedar.Environment) error {
-	conf, session, err := cedar.GetSessionWithConfig(env)
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	defer session.Close()
-	return errors.WithStack(session.DB(conf.DatabaseName).DropDatabase())
-}
 
 func TestFTDCRollupsJob(t *testing.T) {
 	if runtime.GOOS == "darwin" && os.Getenv("EVR_TASK_ID") != "" {
