+++ units/units_test.go
@@ -1,13 +1,42 @@
 package units
 
 import (
+	"context"
 	"testing"
+	"time"
 
+	"github.com/evergreen-ci/cedar"
 	"github.com/mongodb/amboy"
 	"github.com/mongodb/amboy/registry"
 	"github.com/mongodb/grip"
+	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 )
+
+const testDBName = "cedar_test_units"
+
+func init() {
+	env, err := cedar.NewEnvironment(context.Background(), testDBName, &cedar.Configuration{
+		MongoDBURI:    "mongodb://localhost:27017",
+		DatabaseName:  testDBName,
+		SocketTimeout: time.Minute,
+		NumWorkers:    2,
+	})
+	if err != nil {
+		panic(err)
+	}
+
+	cedar.SetEnvironment(env)
+}
+
+func tearDownEnv(env cedar.Environment) error {
+	conf, session, err := cedar.GetSessionWithConfig(env)
+	if err != nil {
+		return errors.WithStack(err)
+	}
+	defer session.Close()
+	return errors.WithStack(session.DB(conf.DatabaseName).DropDatabase())
+}
 
 func TestAllRegisteredUnitsAreRemoteSafe(t *testing.T) {
 	assert := assert.New(t)
